*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@4582 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/modules/sr810.c
@@ -2106,7 +2106,8 @@ static double sr810_get_auto_data( int type )
 			delta_time = new_time - sr810.auto_time;
 
 			if ( delta_time < st_list[ sr810.st_index ] )
-				fsc2_usleep( floor( ( st_list[ sr810.st_index ] - delta_time )
+				fsc2_usleep( ( unsigned long )
+							 floor( ( st_list[ sr810.st_index ] - delta_time )
 									* 1000000.0 ), UNSET );
 			new_try = UNSET;
 		}
